Unify DeepSeek V4 prefill layer EP2 entry
